Allow explicit sample DSL configuration together with functional tests
