+++ src/functTest/groovy/org/gradle/samples/BasicConventionalGroovyDslSampleFunctionalTest.groovy
@@ -1,0 +1,53 @@
+package org.gradle.samples
+
+import org.gradle.testkit.runner.BuildResult
+
+class BasicConventionalGroovyDslSampleFunctionalTest extends AbstractBasicSampleFunctionalTest {
+    @Override
+    protected void makeSingleProject() {
+        buildFile << """
+            plugins {
+                id 'org.gradle.samples'
+            }
+
+            samples {
+                create("demo") {
+                    sampleDir = file('src')
+                }
+            }
+        """
+    }
+
+    @Override
+    protected void writeSampleUnderTest() {
+        temporaryFolder.newFolder("src")
+        temporaryFolder.newFile("src/README.adoc") << """
+= Demo Sample
+
+Some doc
+
+ifndef::env-github[]
+- link:{zip-base-file-name}-groovy-dsl.zip[Download Groovy DSL ZIP]
+endif::[]
+"""
+        writeGroovyDslSample("src")
+    }
+
+    @Override
+    protected void assertSampleTasksExecutedAndNotSkipped(BuildResult result) {
+        assertOnlyGroovyDslTasksExecutedAndNotSkipped(result);
+    }
+
+    @Override
+    protected void assertSampleIndexContainsLinkToSampleArchives() {
+        def sampleIndexFile = new File(projectDir, "build/gradle-samples/demo/index.html")
+        assert sampleIndexFile.exists()
+        assert sampleIndexFile.text.contains('<a href="demo-groovy-dsl.zip">')
+    }
+
+    @Override
+    protected void assertZipsHasContent() {
+        assertZipHasContent(groovyDslZipFile, "gradlew", "gradlew.bat", "gradle/wrapper/gradle-wrapper.properties", "gradle/wrapper/gradle-wrapper.jar", "README.adoc", "build.gradle", "settings.gradle")
+        assert !kotlinDslZipFile.exists()
+    }
+}
