+++ src/main/groovy/org/gradle/samples/Sample.java
@@ -11,4 +11,8 @@ public interface Sample extends Named {
     Property<String> getGradleVersion();
 
     ConfigurableFileCollection getArchiveContent();
+
+    void withGroovyDsl();
+
+    void withKotlinDsl();
 }
